% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{addFakeEntriesForDaysWithoutData}
\alias{addFakeEntriesForDaysWithoutData}
\title{Add Fake Entries for Days Without Data}
\usage{
addFakeEntriesForDaysWithoutData(dataFrame, parameterName = "Q")
}
\arguments{
\item{dataFrame}{data frame with columns \emph{day}, \emph{parName},
\emph{parVal}}

\item{parameterName}{name of parameter, default: \code{"Q"}}
}
\description{
Add Fake Entries for Days Without Data
}
