% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{addStatisticsToEvents}
\alias{addStatisticsToEvents}
\title{Add Statistics to Events}
\usage{
addStatisticsToEvents(events, hydraulicData)
}
\arguments{
\item{events}{list with elements \code{hydraulic} and \code{merged} each
of which is a data frame containing event information}

\item{hydraulicData}{passed to \code{\link{getStatisticsByEvent}}}
}
\description{
Add Statistics to Events
}
