% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{addZoomToHistory}
\alias{addZoomToHistory}
\title{Add Zoom to History}
\usage{
addZoomToHistory(zoomHistory, i, j)
}
\arguments{
\item{zoomHistory}{list of (i, j) pairs, storing the history of zoom settings}

\item{i}{index in range of slider "left"}

\item{j}{index in range of slider "right"}
}
\description{
Add Zoom to History
}
