% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{appendColumn_samplesOk}
\alias{appendColumn_samplesOk}
\title{Append Column "samplesOk"}
\usage{
appendColumn_samplesOk(samplingEvents, bottleEvents, successWord = "SUCCESS",
  columnName = "samplesOk")
}
\arguments{
\item{samplingEvents}{data frame with columns \code{samplerFile}, 
\code{bottle}, \code{result}}

\item{bottleEvents}{bottle events (data frame)}

\item{successWord}{word indicating a successful sampling in column
\code{result} of \code{samplingEvents}}

\item{columnName}{name of appended column}
}
\description{
Append Column "samplesOk"
}
