% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{appendInterpolColumns}
\alias{appendInterpolColumns}
\title{Append Interpol Columns}
\usage{
appendInterpolColumns(hydraulicData, settings, columnQraw = "Q.raw",
  columnQ = "Q", columnQInterpol = "Q.interpol", columnHraw = "H.raw",
  columnH = "H", columnHInterpol = "H.interpol")
}
\arguments{
\item{hydraulicData}{data frame containing hydraulic data}

\item{settings}{list of settings (not used!)}

\item{columnQraw}{name of column containing raw discharge}

\item{columnQ}{name of column containing (valid) discharge}

\item{columnQInterpol}{name of column to be added containing interpolated
discharges}

\item{columnHraw}{name of column containing raw water levels}

\item{columnH}{name of column containing (valid) water levels}

\item{columnHInterpol}{name of column to be added containing interpolated
water levels}
}
\value{
data frame with columns \emph{H.interpol}, \emph{Q.interpol}
  appended, containing only the interpolated values and NA for times when H
  (or Q, respectively) was already available in column \emph{H.raw} (or 
  \emph{Q.raw}, respectively)
}
\description{
Append Interpol Columns
}
