% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{apply_H_threshold}
\alias{apply_H_threshold}
\title{Apply H Threshold}
\usage{
apply_H_threshold(dat.raw, settings)
}
\arguments{
\item{dat.raw}{data frame with column \emph{H}}

\item{settings}{list as returned by \code{\link{configure}} with list element
\emph{Hthresholds}}
}
\description{
Apply H threshold given in settings to H in \emph{dat.raw}
}
