% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{availableAutoSamplerFiles}
\alias{availableAutoSamplerFiles}
\title{Available "sample_log"-Files}
\usage{
availableAutoSamplerFiles(sampleDirectory = getOrCreatePath("SAMPLE_DIR",
  dictionary), pattern = do_resolve("SAMPLE_CSV_PATTERN", dictionary),
  dictionary = NULL, warn = TRUE)
}
\arguments{
\item{sampleDirectory}{directory in which to look for sample files. By
default the directory is looked up in the \emph{dictionary} at keyword:
SAMPLE_DIR}

\item{pattern}{file name pattern to which file names are matched. By default
pattern is looked up in the \emph{dictionary} at keyword:
SAMPLE_CSV_PATTERN}

\item{dictionary}{dictionary (list) with elements \emph{SAMPLE_DIR} and 
\emph{SAMPLE_CSV_PATTERN}}

\item{warn}{if TRUE, a warning is given if there are no sample files}
}
\description{
Available "sample_log"-Files
}
