% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{availableAutoSamplerFiles2}
\alias{availableAutoSamplerFiles2}
\title{Available Auto Sampler Files 2}
\usage{
availableAutoSamplerFiles2(rootDirectory, station, dictionaryFile)
}
\arguments{
\item{rootDirectory}{passed as \code{RAW_DIR} to \code{\link{pathDictionary}}}

\item{station}{passed as \code{STATION} to \code{\link{pathDictionary}}}

\item{dictionaryFile}{passed to \code{\link{pathDictionary}}}
}
\description{
Available Auto Sampler Files 2
}
