% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{bottleEventsToSamplerEvents}
\alias{bottleEventsToSamplerEvents}
\title{Bottle Events to Sampler Events}
\usage{
bottleEventsToSamplerEvents(bottleEvents, signalWidth)
}
\arguments{
\item{bottleEvents}{data frame with columns \emph{samplerFile}}

\item{signalWidth}{passed to \code{\link[kwb.event]{toEvents}}}
}
\description{
Bottle Events to Sampler Events
}
