% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_sample.R
\name{calculateVolumeCompositeSample}
\alias{calculateVolumeCompositeSample}
\title{Calculate Volume Composite Sample}
\usage{
calculateVolumeCompositeSample(hydraulicSubset, settings,
  show.all.bottles = FALSE)
}
\arguments{
\item{hydraulicSubset}{data frame with columns \emph{DateTime}, \emph{Q}, 
\emph{bottle}}

\item{settings}{settings as returned by \code{\link{configure}}}

\item{show.all.bottles}{if \code{TRUE}, the volume table is shown not only
for the selected bottles but also for all bottles (inclusive discarded
ones). Default: \code{FALSE}}
}
\description{
Calculate Volume Composite Sample
}
