% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dateToDateStringInPath}
\alias{dateToDateStringInPath}
\title{Date to Date String in Path}
\usage{
dateToDateStringInPath(x)
}
\arguments{
\item{x}{date or time object passed to \code{\link[kwb.datetime]{hsDateStr}}}
}
\description{
Date to Date String in Path
}
