% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rain.R
\name{.plotRainData}
\alias{.plotRainData}
\title{Plot Rain Data}
\usage{
.plotRainData(rainData, gauges, xlim, innerMargins, eventAndStat = NULL)
}
\arguments{
\item{rainData}{data frame with columns \emph{DateTime}}

\item{gauges}{character vector of gauge names}

\item{xlim}{vector of two POSIXct determining the x limits}

\item{innerMargins}{passed to \code{\link[kwb.plot]{plotRain}}}

\item{eventAndStat}{if not NULL (default) this must be a one row data frame
with columns ...}
}
\description{
Plot Rain Data
}
