% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.toIntervalList}
\alias{.toIntervalList}
\title{Character Vectors to List of "lubridate" Intervals}
\usage{
.toIntervalList(from = NULL, to = NULL, tzone = "UTC")
}
\arguments{
\item{from}{vector of timestamps (vector of character will be converted to
vector of POSIXct using \code{kwb.datetime::stringToPosix}) representing
the starts of the intervals}

\item{to}{vector of timestamps (vector of character will be converted to
vector of POSIXct using \code{kwb.datetime::stringToPosix}) representing
the ends of the intervals}

\item{tzone}{passed to \code{kwb.datetime::stringToPosix} and 
\code{lubridate::interval}}
}
\description{
Converts two character vectors representing the beginning and end timestamps 
of time intervals into a list of lubridate interval objects created by 
\code{interval}
}
