% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{filterForRelevantBottles}
\alias{filterForRelevantBottles}
\title{Filter Data Frame by Bottle Number}
\usage{
filterForRelevantBottles(sampleDataExtended, bottlesToConsider)
}
\arguments{
\item{sampleDataExtended}{dataFrame with column \emph{bottle}}

\item{bottlesToConsider}{vector of bottle numbers to be considered}
}
\value{
data frame with rows in which column \emph{bottle} is one of the
  numbers given in \emph{bottlesToConsider}
}
\description{
Filter Data Frame by Bottle Number
}
