% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{filterSampleEventsForFilename}
\alias{filterSampleEventsForFilename}
\title{Filter Sample Events for Filename}
\usage{
filterSampleEventsForFilename(events, fileName)
}
\arguments{
\item{events}{list with required elements \emph{samplingEvents},
\emph{bottleEvents}, \emph{samplerEvents}}

\item{fileName}{file name to be filtered for in column \emph{samplerFile} of
each of the data frames \emph{events$samplingEvents},
\emph{events$bottleEvents} and \emph{events$samplerEvents}}
}
\value{
list with elements \emph{samplingEvents}, \emph{bottleEvents}, 
  \emph{samplerEvents}
}
\description{
Filter Sample Events for Filename
}
