% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{formatEvent}
\alias{formatEvent}
\title{Format Event}
\usage{
formatEvent(event, eventNumber = 1, precisionLevel = NULL)
}
\arguments{
\item{event}{one row of event data frame as returned by 
\code{\link[kwb.event]{hsEvents}}}

\item{eventNumber}{number of the event}

\item{precisionLevel}{1 (less precise) or 2 (more precise)}
}
\description{
Format Event
}
