% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{formatEventStatistics}
\alias{formatEventStatistics}
\title{Format Event Statistics}
\usage{
formatEventStatistics(eventStatistics, precisionLevel = NULL)
}
\arguments{
\item{eventStatistics}{date frame with columns \emph{V.m3}, \emph{H.max},
\emph{Q.max}, \emph{Q.raw.na},}

\item{precisionLevel}{number of digits after decimal point}
}
\description{
Format Event Statistics
}
