% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{formatEventStatisticsTable}
\alias{formatEventStatisticsTable}
\title{Format Event Statistics Table}
\usage{
formatEventStatisticsTable(eventStatisticsExtended, precisionLevel = NULL)
}
\arguments{
\item{eventStatisticsExtended}{list of event properties}

\item{precisionLevel}{1 (less precise) or 2 (more precise)}
}
\description{
Format Event Statistics Table
}
