% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{formatSettings}
\alias{formatSettings}
\title{Format Settings}
\usage{
formatSettings(settings, settingNames = names(settings),
  do.stop = FALSE)
}
\arguments{
\item{settings}{list of settings}

\item{settingNames}{names of the settings, by default: 
\code{names(settings)}}

\item{do.stop}{passed to \code{kwb.monitoring:::get_H_threshold}, 
\code{kwb.monitoring:::get_Q_threshold}, 
\code{kwb.monitoring:::get_V_threshold}}
}
\description{
Format Settings
}
