% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{getAllSamplerEvents}
\alias{getAllSamplerEvents}
\title{Get All Sampler Events}
\usage{
getAllSamplerEvents(rootDirectory, dictionaryFile, stations,
  FUN.readSamplerFile, bottlesToConsider = NA,
  sampleEventSeparationTime = 3600, method = "centre",
  signalWidth = 60)
}
\arguments{
\item{rootDirectory}{passed to \code{availableAutoSamplerFiles2}}

\item{dictionaryFile}{passed to \code{availableAutoSamplerFiles2}}

\item{stations}{vector of monitoring station names, each of which is passed
to \code{availableAutoSamplerFiles2}}

\item{FUN.readSamplerFile}{passed to \code{getAllSamplerEventsFromFiles}}

\item{bottlesToConsider}{passed to \code{getAllSamplerEventsFromFiles}}

\item{sampleEventSeparationTime}{passed to
\code{getAllSamplerEventsFromFiles}}

\item{method}{passed to \code{getAllSamplerEventsFromFiles}}

\item{signalWidth}{passed to \code{getAllSamplerEventsFromFiles}}
}
\description{
Get All Sampler Events
}
