% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{getAllTypesOfEvents}
\alias{getAllTypesOfEvents}
\title{Get All Types of Events}
\usage{
getAllTypesOfEvents(hydraulicData = NULL, settings, FUN.readSamplerFile)
}
\arguments{
\item{hydraulicData}{data frame containing hydraulic data that is passed to
\code{\link{getHydraulicEvents}}}

\item{settings}{list of settings, passed to \code{\link{getSamplerEvents}} 
and containing element \code{tstep.fill.s}}

\item{FUN.readSamplerFile}{e.g. kwb.ogre::readOgreSamplerFileByName or 
kwb.dswt::readDswtSamplerFileByName}
}
\description{
Get All Types of Events
}
