% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getFunctionValueOrDefault2}
\alias{getFunctionValueOrDefault2}
\title{Get Function Value or Default 2}
\usage{
getFunctionValueOrDefault2(values, FUN, default, timestamps = NULL,
  columnName = "")
}
\arguments{
\item{values}{passed to \code{\link[kwb.utils]{getFunctionValueOrDefault}}}

\item{FUN}{passed to \code{\link[kwb.utils]{getFunctionValueOrDefault}}}

\item{default}{passed to \code{\link[kwb.utils]{getFunctionValueOrDefault}}}

\item{timestamps}{vector of timestamps (used in warning message)}

\item{columnName}{column name (used in warning message)}
}
\description{
Get Function Value or Default 2
}
