% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{getHydraulicEvents}
\alias{getHydraulicEvents}
\title{Get Hydraulic Events}
\usage{
getHydraulicEvents(hydraulicData, settings,
  eventSettings = settings$event[[settings$station]], columnQ = "Q",
  columnH = "H")
}
\arguments{
\item{hydraulicData}{data frame with columns ...}

\item{settings}{settings as returned by ...}

\item{eventSettings}{default: \code{settings$event[[settings$station]]}}

\item{columnQ}{name of column containing water flows. Default: "Q"}

\item{columnH}{name of column containing water levels. Default: "H"}
}
\description{
Get Hydraulic Events
}
