% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{getIndicesWithinEvents}
\alias{getIndicesWithinEvents}
\title{Get Indices Within Events}
\usage{
getIndicesWithinEvents(hydraulicData, eventSettings = NULL, thresholds = c(H
  = NA, Q = NA), columns = c(H = "H", Q = "Q"))
}
\arguments{
\item{hydraulicData}{data frame with a columns \emph{DateTime} and two
columns named as given in \emph{columnH}, \emph{columnQ}}

\item{eventSettings}{data frame with columns \emph{tBeg}, \emph{tEnd} (begin
and end of period in which thresholds are valid), \emph{Hthreshold},
\emph{Qthreshold}}

\item{thresholds}{default H and Q thresholds to be applied for time intervals
for which no special thresholds are defined}

\item{columns}{optional. Named vector of character with column names for
\code{H} and \code{Q}}
}
\description{
Get Indices Within Events
}
