% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{getMergedEvents}
\alias{getMergedEvents}
\title{Merge hydraulic events and sampler events}
\usage{
getMergedEvents(hydraulicEvents, samplerEvents, signalWidth)
}
\arguments{
\item{hydraulicEvents}{data frame containing information on hydraulic events}

\item{samplerEvents}{data frame containing information on sampler events}

\item{signalWidth}{passed to \code{\link[kwb.event]{toEvents}}}
}
\description{
Merge hydraulic events and sampler events
}
