% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{getModelFromFile}
\alias{getModelFromFile}
\title{Read Regression Model from Text File}
\usage{
getModelFromFile(modelFile, warn = TRUE)
}
\arguments{
\item{modelFile}{path to .RData file containing the model}

\item{warn}{if \code{TRUE} (the default), a warning is given if the model
file does not exist}
}
\description{
Read Regression Model from Text File
}
