% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getPredictionOfQ}
\alias{getPredictionOfQ}
\title{Get Prediction of Q}
\usage{
getPredictionOfQ(hydraulicData, regressionModels = NULL, modelDir,
  columns = c(DateTime = "DateTime", H = "H", Q.raw = "Q.raw"),
  only.if.na = TRUE)
}
\arguments{
\item{hydraulicData}{data frame with columns as named in \code{columns}}

\item{regressionModels}{data frame with character columns \emph{from},
\emph{to} (or POSIXct columns \emph{tBeg} and \emph{tEnd}) and character
columns \emph{modelFile} determining the time intervals to which the
different correlation models are assigned.}

\item{modelDir}{full path to directory where the model files (of the current
station) are stored}

\item{columns}{names of columns containing Date and Time, water levels and 
water flows}

\item{only.if.na}{logical. Not used!}
}
\description{
Get Prediction of Q
}
