% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{getSampleInformation}
\alias{getSampleInformation}
\title{Get Sample Information}
\usage{
getSampleInformation(dictionary = settings$dictionary,
  sampleEventSeparationTime = settings$sampleEventSeparationTime,
  sampleEventIndices = -1, bottlesToConsider = settings$bottlesToConsider,
  method = settings$sampleEventMethod, FUN.readSamplerFile, settings = NULL,
  signalWidth = 1)
}
\arguments{
\item{dictionary}{default: settings$dictionary}

\item{sampleEventSeparationTime}{default: settings$sampleEventSeparationTime}

\item{sampleEventIndices}{default: -1}

\item{bottlesToConsider}{default: settings$bottlesToConsider}

\item{method}{one of the methods supported by
\code{\link{sampleDataToSamplingEvents}}. default:
settings$sampleEventMethod}

\item{FUN.readSamplerFile}{function to be used to read an auto sampler file}

\item{settings}{list of settings, as returned by \code{\link{configure}}}

\item{signalWidth}{signal width in seconds. Default: 1}
}
\value{
list with elements \emph{sampleTimes}, \emph{samplingEvents},
  \emph{bottle events} (data frame with columns...)
}
\description{
Get Sample Information
}
