% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{getSamplerEvents}
\alias{getSamplerEvents}
\title{Get Sampler Events}
\usage{
getSamplerEvents(settings, FUN.readSamplerFile)
}
\arguments{
\item{settings}{passed to \code{\link{getSampleInformation}}}

\item{FUN.readSamplerFile}{passed to \code{\link{getSampleInformation}}}
}
\description{
Get sample event information from all available auto sampler files
}
