% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{getStatisticsByDay}
\alias{getStatisticsByDay}
\title{Get Statistics by Day}
\usage{
getStatisticsByDay(dataFrame)
}
\arguments{
\item{dataFrame}{data frame with columns \emph{parName}, \emph{parVal},
\emph{day}}
}
\value{
data frame with additional columns \emph{info} ("day: H max = ... cm,
  Q max = ... L/s")
}
\description{
Get Statistics by Day
}
