% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{getStatisticsByEvent}
\alias{getStatisticsByEvent}
\title{Get Statistics by Event}
\usage{
getStatisticsByEvent(hydraulicData, events)
}
\arguments{
\item{hydraulicData}{data frame containing hydraulic data}

\item{events}{data frame containing information on events, passed to 
\code{\link[kwb.event]{hsEventNumber}}}
}
\description{
Get Statistics by Event
}
