% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{indicesInIntervals}
\alias{indicesInIntervals}
\title{Indices of Rows Belonging to Time Intervals}
\usage{
indicesInIntervals(timestamps, intervals)
}
\arguments{
\item{timestamps}{vector of POSIXct timestamps}

\item{intervals}{vector of Interval objects as returned by
lubridate::interval}
}
\value{
indices of elements in \emph{timestamps} that belong to the given
  time \emph{intervals}
}
\description{
Indices of Rows Belonging to Time Intervals
}
