% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{mergeParallelRainEventStat}
\alias{mergeParallelRainEventStat}
\title{Merge Parallel Rain Event Statistics}
\usage{
mergeParallelRainEventStat(hydraulicEvents, rainEvents, rainData,
  seriesName, offset = 0, plot.merged.event.info = TRUE)
}
\arguments{
\item{hydraulicEvents}{data frame containing information on hydraulic events,
with columns \code{eventNumber, tBeg}}

\item{rainEvents}{list of data frames containing information on rain events,
with one list entry per rain gauge of which one is named as given in
\code{seriesName}}

\item{rainData}{passed to \code{\link[kwb.event]{getEventStatistics}}}

\item{seriesName}{name of rain gauge}

\item{offset}{time in seconds by which tBeg is shifted backwards}

\item{plot.merged.event.info}{if \code{TRUE} (default), 
\code{\link[kwb.event]{plotMergedEventInfoForValidation}} is called}
}
\description{
Merge Parallel Rain Event Statistics
}
