% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{pathDictionary}
\alias{pathDictionary}
\title{Path Dictionary}
\usage{
pathDictionary(dictionaryFile, RAW_DIR = settings$rawdir,
  STATION = settings$station, settings = NULL)
}
\arguments{
\item{dictionaryFile}{full path to file defining a dictionary and being 
read with \code{\link[kwb.utils]{readDictionary}}}

\item{RAW_DIR}{value for placeholder of the same name in the dictionary}

\item{STATION}{value for placeholder of the same name in the dictionary}

\item{settings}{optional. List with elements \code{rawdir, station} from 
which to take values to be used for \code{RAW_DIR, STATION}}
}
\description{
Read dictionary from file and set RAW_DIR and STATION
}
