% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_H_columns}
\alias{plot_H_columns}
\title{Plot H Columns}
\usage{
plot_H_columns(hydraulicData, h.threshold = 0,
  time.dependent.thresholds = NULL, xlim = NULL, ylim = NULL,
  innerMargins = .defaultInnerMargins())
}
\arguments{
\item{hydraulicData}{data frame with columns \emph{DateTime}, \emph{H},
\emph{H.interpol}}

\item{h.threshold}{H threshold at which a horizontal line is to be drawn
(default: 0)}

\item{time.dependent.thresholds}{passed to 
\code{kwb.monitoring:::.drawAdditionalThresholdsIfApplicable}}

\item{xlim}{passed to \code{\link[kwb.plot]{plot_variable}}}

\item{ylim}{passed to \code{\link[kwb.plot]{plot_variable}}}

\item{innerMargins}{passed to \code{\link[kwb.plot]{plot_variable}}}
}
\description{
Plot H Columns
}
