% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydraulic_event.R
\name{plot_hydraulic_event}
\alias{plot_hydraulic_event}
\title{Plot Hydraulic Event}
\usage{
plot_hydraulic_event(hydraulicData, settings, eventAndStat,
  sampleInformation = NULL, ylim.Q = NULL, rainData = NULL,
  gauges = NULL, left = 0.1, right = 0.1, innerMargins.HQ = c(0.2,
  left, 0.1, right), innerMargins.rain = c(0, left, 0.2, right),
  dbg = FALSE)
}
\arguments{
\item{hydraulicData}{data frame with columns \emph{DateTime}, \emph{H},
\emph{Q}, \emph{Q.raw}, \emph{Q.raw.signal}, \emph{Q.interpol}}

\item{settings}{list of settings containing additional information
and passed to other functions}

\item{eventAndStat}{passed to \code{kwb.monitoring:::.plotRainData}}

\item{sampleInformation}{passed to \code{kwb.monitoring:::.partialPlot_H}}

\item{ylim.Q}{passed to \code{kwb.monitoring:::.partialPlot_Q}}

\item{rainData}{optional. Data frame containing rain data}

\item{gauges}{passed to \code{kwb.monitoring:::.plotRainData} if 
\code{rainData} is given}

\item{left}{fraction of event length by which xlim is extended to the left}

\item{right}{fraction of event length by which xlim is extended to the right}

\item{innerMargins.HQ}{"inner margins" of H and Q plots. Default: c(0.2,
left, 0.1, right)}

\item{innerMargins.rain}{"inner margins" of rain plots. Default: c(0, left,
0.2, right)}

\item{dbg}{logical. If \code{TRUE} \code{eventAndStat} is printed.}
}
\description{
Plot Hydraulic Event
}
