% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hydraulic_events}
\alias{plot_hydraulic_events}
\title{Plot Hydraulic Events}
\usage{
plot_hydraulic_events(hydraulicData, settings, eventsAndStat, to.pdf = FALSE,
  rainData = NULL, gauges = NULL, landscape = TRUE,
  plot.event.overview = TRUE,
  FUN.plot_hydraulic_event = kwb.monitoring::plot_hydraulic_event, ...)
}
\arguments{
\item{hydraulicData}{data frame with column \code{DateTime}, passed to
the function given in \code{FUN.plot_hydraulic_event}}

\item{settings}{list of settings (e.g. \code{dictionary}), passed to 
\code{\link{plotEventDistribution}} if \code{plot.event.overview} is 
\code{TRUE}}

\item{eventsAndStat}{data frame containing event information}

\item{to.pdf}{if \code{TRUE}, graphical output goes to a temporary pdf file}

\item{rainData}{passed to the function given in 
\emph{FUN.plot_hydraulic_event}}

\item{gauges}{passed to the function given in 
\emph{FUN.plot_hydraulic_event}}

\item{landscape}{orientation of pages in PDF file if \code{to.pdf} is 
\code{TRUE}}

\item{plot.event.overview}{if \code{TRUE}, 
\code{\link{plotEventDistribution}} is called}

\item{FUN.plot_hydraulic_event}{function to be called to plot one event}

\item{\dots}{arguments passed to the function given in 
\emph{FUN.plot_hydraulic_event}}
}
\description{
Plot Hydraulic Events
}
