% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sampled_event}
\alias{plot_sampled_event}
\title{Plot Sampled Event}
\usage{
plot_sampled_event(hydraulicData, settings, sampleInformation = NULL,
  mergedEventAndStat, volumeCompositeSample = NULL, to.pdf = FALSE,
  interpolate = TRUE, ...)
}
\arguments{
\item{hydraulicData}{data frame with columns \emph{DateTime}, \emph{H},
\emph{Q}, \emph{Q.raw}, \emph{Q.raw.signal}, \emph{Q.interpol}}

\item{settings}{list with elements \code{dictionary, station, precisionLevel, 
bottlesToDiscard, bottlesToConsider} and being passed to other functions}

\item{sampleInformation}{passed to \code{kwb.monitoring:::.partialPlot_Q} and
\code{kwb.monitoring:::.partialPlot_H}}

\item{mergedEventAndStat}{list with event information such as \code{tBeg},
\code{tEnd}, \code{V.m3} and passed to 
  \code{kwb.monitoring:::formatEventRelation}}

\item{volumeCompositeSample}{list with elements \code{bottle}, {V.bottle.mL}, 
\code{V}, passed to \code{kwb.monitoring:::formatVolumeCompositeSample}}

\item{to.pdf}{logical. If \code{TRUE}, output goes to a PDF file}

\item{interpolate}{passed to \code{kwb.monitoring:::.partialPlot_Q}}

\item{\dots}{further arguments given to}
}
\description{
Plot Sampled Event
}
