% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_high_level.R
\name{plotEventDistribution}
\alias{plotEventDistribution}
\title{Plot Event Distribution}
\usage{
plotEventDistribution(eventsAndStat, settings)
}
\arguments{
\item{eventsAndStat}{data frame containing event information}

\item{settings}{list of settings, at least with elements 
\code{station, precisionLevel}}
}
\description{
Plot Event Distribution
}
