% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotEventOverview}
\alias{plotEventOverview}
\title{Gantt-Plots for Event Overview}
\usage{
plotEventOverview(events, settings, dbg = FALSE)
}
\arguments{
\item{events}{list with elements \emph{hydraulic}, \emph{sample},
\emph{merged} each of which is a data frame with columns \emph{tBeg},
\emph{tEnd}}

\item{settings}{list with elements \emph{evtSepTime}, \emph{station},
\emph{Hthresholds}}

\item{dbg}{logical. If \code{TRUE}, the x axis coordinates are printed}
}
\description{
Gantt-Plots for Event Overview
}
