% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{plotOverview}
\alias{plotOverview}
\title{Plot Overview}
\usage{
plotOverview(dat, station, Qmax = NULL, Hmax = NULL)
}
\arguments{
\item{dat}{data frame containing the data to be plotted}

\item{station}{name of monitoring station, used in plot title}

\item{Qmax}{passed to \code{kwb.monitoring:::plotOverview_byDay}}

\item{Hmax}{passed to \code{kwb.monitoring:::plotOverview_byDay}}
}
\description{
Plot Overview
}
