% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_information.R
\name{plotSampleInformation}
\alias{plotSampleInformation}
\title{Plot Sample Information}
\usage{
plotSampleInformation(sampleInformation, add = FALSE,
  xlim = toUTC(range(c(sampleInformation$samplingEvents$tBeg,
  sampleInformation$samplingEvents$tEnd))), ylim = c(-2, 7), main = NA,
  cex.legend = 0.6, density = 0, plotSampleIntervals = TRUE,
  maxSamplesOk = NULL, plotSamplingPoints = plotSampleIntervals)
}
\arguments{
\item{sampleInformation}{list with elements \emph{samplingEvents},
\emph{bottleEvents}}

\item{add}{passed to \code{\link[kwb.event]{ganttPlotEvents}}}

\item{xlim}{passed to \code{\link[kwb.event]{ganttPlotEvents}}}

\item{ylim}{passed to \code{\link[kwb.event]{ganttPlotEvents}}}

\item{main}{plot title}

\item{cex.legend}{passed to  \code{\link{addSampleTimesToPlot}}}

\item{density}{passed to \code{\link[kwb.event]{ganttPlotEvents}}}

\item{plotSampleIntervals}{logical. If \code{TRUE} 
\code{\link[kwb.event]{ganttPlotEvents}} is called for 
  \code{sampleInformation$samplingEvents}}

\item{maxSamplesOk}{maximum number of valid samples. This value is used to
calculate an y coordinate}

\item{plotSamplingPoints}{= logical. If \code{TRUE} 
\code{\link{addSampleTimesToPlot}} is called}
}
\description{
plotSampleInformation. TODO: simplify interface, e.g. 
  plotSampleInformation(getSampleInformation(getSampleFiles()[1]))
}
