% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTotalDischargeVersusRainProperties}
\alias{plotTotalDischargeVersusRainProperties}
\title{Plot Total Discharge Versus Rain Properties}
\usage{
plotTotalDischargeVersusRainProperties(hydraulicEvents, stations, main,
  settings, statistics = "sum", to.pdf = FALSE)
}
\arguments{
\item{hydraulicEvents}{data frame containing information on hydraulic events}

\item{stations}{vector of monitoring station names}

\item{main}{plot title}

\item{settings}{list of settings}

\item{statistics}{one of \code{c("sum", "mean", "max")}}

\item{to.pdf}{logial. If \code{TRUE}, the output goes into a PDF file}
}
\description{
Plot Total Discharge Versus Rain Properties
}
