% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{printSampleInformation}
\alias{printSampleInformation}
\title{Print Sample Information}
\usage{
printSampleInformation(sampleInformation)
}
\arguments{
\item{sampleInformation}{list with elements \code{sampleData, samplingEvents, 
bottleEvents}}
}
\description{
Print Sample Information
}
