% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{rainGaugesNearStation}
\alias{rainGaugesNearStation}
\title{Rain Gauges Near to Monitoring Sites}
\usage{
rainGaugesNearStation(station = NULL)
}
\arguments{
\item{station}{name of monitoring station in KWB project OGRE or DSWT}
}
\value{
list (one list element per monitoring site) of character vectors 
  representing rain gauge names
}
\description{
Rain Gauges Near to Monitoring Sites
}
