% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{readAndJoinSamplerFiles}
\alias{readAndJoinSamplerFiles}
\title{Read Multiple Auto-Sampler Files}
\usage{
readAndJoinSamplerFiles(samplerFiles, FUN.readSamplerFile,
  bottlesToConsider = NA)
}
\arguments{
\item{samplerFiles}{vector of paths to sampler files}

\item{FUN.readSamplerFile}{function to be used for reading the sampler file}

\item{bottlesToConsider}{vector of bottle numbers to consider. Defaults to NA
meaning that information on all bottles are to be returened.}
}
\value{
data frame with columns \emph{file}, \emph{myDateTime},
  \emph{sample}, \emph{bottle}, \emph{volume}, \emph{unit}, \emph{result}
}
\description{
Read Multiple Auto-Sampler Files
}
