% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove Duplicates}
\usage{
removeDuplicates(hydraulicData, timeColumnName = "DateTime",
  keep.first = TRUE)
}
\arguments{
\item{hydraulicData}{data frame with date and time column as named in 
\code{timeColumnName}}

\item{timeColumnName}{name of date and time column in \code{hydraulicData}}

\item{keep.first}{logical. If \code{TRUE} (\code{FALSE} is not implemented!)
only the first rows of sets of rows with duplicated time stamps are kept.}
}
\description{
Remove Duplicates
}
