% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{removeIntervals}
\alias{removeIntervals}
\title{Remove Intervals}
\usage{
removeIntervals(dataFrame, intervals,
  dateTimeColumn = names(kwb.utils::posixColumnAtPosition(dataFrame))[1])
}
\arguments{
\item{dataFrame}{data frame with a column as named in \code{dateTimeColumn}}

\item{intervals}{vector of Interval objects as returned by
lubridate::interval, passed to \code{\link{indicesInIntervals}}}

\item{dateTimeColumn}{name of date and time column in \code{dataFrame}}
}
\description{
Remove Intervals
}
