% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{removeZoomFromHistory}
\alias{removeZoomFromHistory}
\title{Remove Zoom from History}
\usage{
removeZoomFromHistory(zoomHistory)
}
\arguments{
\item{zoomHistory}{list of (i, j) pairs, storing the history of zoom settings}
}
\description{
Remove Zoom from History
}
