% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sampleDataToSamplingEvents}
\alias{sampleDataToSamplingEvents}
\title{Sample Data to Sampling Events}
\usage{
sampleDataToSamplingEvents(sampleData, method = "centre",
  default.interval.width.s = 600, signalWidth = 1)
}
\arguments{
\item{sampleData}{sample data as returned by
\code{\link{readAndJoinSamplerFiles}}}

\item{method}{one of c("left", "right", "centre")}

\item{default.interval.width.s}{default interval width in seconds}

\item{signalWidth}{interval length in seconds that a sampling action is
assumed to represent. It is only used to calculate the duration D of the
time interval between two samplings at t1 and t2: D = t2 - t1 + signalWidth}
}
\value{
data frame with columns \emph{tBeg}, \emph{tEnd}, ...
}
\description{
Sample Data to Sampling Events
}
