% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sampleLogFileToSampleName}
\alias{sampleLogFileToSampleName}
\title{Sample Log File to Sample Name}
\usage{
sampleLogFileToSampleName(sampleFile)
}
\arguments{
\item{sampleFile}{full path to auto sampler file}
}
\description{
Sample Log File to Sample Name
}
