% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{samplingEventsToBottleEvents}
\alias{samplingEventsToBottleEvents}
\title{Sampling Events to Bottle Events}
\usage{
samplingEventsToBottleEvents(samplingEvents, signalWidth = 1)
}
\arguments{
\item{samplingEvents}{data frame with columns \emph{samplerFile},
\emph{bottle}, \emph{tBeg}, \emph{tEnd}}

\item{signalWidth}{passed to \code{\link[kwb.event]{toEvents}}}
}
\value{
data frame with columns \emph{tBeg}, \emph{tEnd}, \emph{dur}, 
  \emph{bottle}, \emph{samplesOk}
}
\description{
Sampling Events to Bottle Events
}
