% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{saveRegressionModel}
\alias{saveRegressionModel}
\title{Save Regression Model}
\usage{
saveRegressionModel(regressionModel, settings = NULL,
  dictionary = kwb.utils::selectElements(settings, "dictionary"),
  sep = kwb.utils::selectElements(settings, "outsep"),
  dec = kwb.utils::selectElements(settings, "outdec"))
}
\arguments{
\item{regressionModel}{model object to be stored}

\item{settings}{optional. List from which to take the arguments if not given}

\item{dictionary}{dictionary (list) containing entries "REGRESSION_MODEL_TXT"
and "REGRESSION_COEFF_TXT"}

\item{sep}{column separator in created file}

\item{dec}{decimal character in created file}
}
\description{
Save Regression Model
}
