% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveSampleInformation}
\alias{saveSampleInformation}
\title{Save Sample Information}
\usage{
saveSampleInformation(sampleInformation, settings, sampleFile)
}
\arguments{
\item{sampleInformation}{list with elements \code{samplingEvents, 
bottleEvents}}

\item{settings}{list with elements \code{dictionary, outsep, outdec}}

\item{sampleFile}{name of auto sampler file, passed to 
\code{\link{sampleLogFileToSampleName}}}
}
\description{
Save Sample Information
}
