% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{selectIntervalsForCorrelation}
\alias{selectIntervalsForCorrelation}
\title{Select Intervals for Correlation}
\usage{
selectIntervalsForCorrelation(dat.all, settings, h.threshold.max = 0.4)
}
\arguments{
\item{dat.all}{data frame containing all relevant data passed to 
\code{kwb.monitoring:::H_above_threshold}}

\item{settings}{list of settings with elements \code{Hthresholds, "station"}
and being passed to \code{kwb.monitoring:::H_above_threshold}}

\item{h.threshold.max}{maximum value for the slider}
}
\description{
Select Intervals for Correlation
}
