% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{setPausesOfMergedEvents}
\alias{setPausesOfMergedEvents}
\title{Update Pauses of Merged Events}
\usage{
setPausesOfMergedEvents(hydraulicEvents, mergedEvents, dbg = FALSE)
}
\arguments{
\item{hydraulicEvents}{data frame containing information on hydraulic events}

\item{mergedEvents}{data frame containing information on "merged" events}

\item{dbg}{logical. If \code{TRUE}, debug messages are shown}
}
\value{
\emph{mergedEvents} with updated pauses \emph{pBefore} and
  \emph{pAfter}
}
\description{
Update Pauses of Merged Events
}
