% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{showOverview}
\alias{showOverview}
\title{Show Overview}
\usage{
showOverview(dat, settings, Qmax = NULL, Hmax = NULL, to.pdf = FALSE,
  save.pdf = FALSE)
}
\arguments{
\item{dat}{data frame with columns \code{DateTime}}

\item{settings}{list with elements \code{dictionary, station}}

\item{Qmax}{passed to \code{\link{plotOverview}}}

\item{Hmax}{passed to \code{\link{plotOverview}}}

\item{to.pdf}{if \code{TRUE}, graphical output goes to a temporary pdf file}

\item{save.pdf}{if \code{TRUE} PDF output is stored to a file stored and 
named according to \code{dictionary$OVERVIEW_HQ_DATA_PDF}}
}
\description{
Show Overview
}
