% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{updateZoomHistory}
\alias{updateZoomHistory}
\title{Update Zoom History}
\usage{
updateZoomHistory(action, zoomHistory, i, j)
}
\arguments{
\item{action}{one of \code{c("zoom.in", "zoom.out")}}

\item{zoomHistory}{list of (i, j) pairs, storing the history of zoom settings}

\item{i}{index in range of slider "left"}

\item{j}{index in range of slider "right"}
}
\description{
Update Zoom History
}
