% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validateAndFillHydraulicData}
\alias{validateAndFillHydraulicData}
\title{Validate and Fill Hydraulic Data}
\usage{
validateAndFillHydraulicData(hydraulicData,
  tstep.fill.s = selectElements(settings, "tstep.fill.s"),
  replaceMissingQMethod = selectElements(settings, "replaceMissingQMethod"),
  regressionModels = selectElements(selectElements(settings, "regression"),
  "models")[[selectElements(settings, "station")]],
  regressionUsage = selectElements(selectElements(settings, "regression"),
  "usage")[[selectElements(settings, "station")]], hydraulicEvents = NULL,
  additionalColumns = NULL, modelDir = getOrCreatePath("REGRESSION_DIR",
  selectElements(settings, "dictionary")), settings = NULL)
}
\arguments{
\item{hydraulicData}{data frame with column \emph{DateTime}, ...}

\item{tstep.fill.s}{target time step in seconds. Time gaps in
\emph{hydraulicData} will be filled with interpolated values}

\item{replaceMissingQMethod}{one of c("interpolate", "predict").
"interpolate": linear interpolation "predict": prediction from water levels
using a saved square regression}

\item{regressionModels}{data frame with character columns \emph{from},
\emph{to} (or POSIXct columns \emph{tBeg} and \emph{tEnd}) and
\emph{modelFile} determining the time intervals to which the different
correlation models are assigned.}

\item{regressionUsage}{data frame with character columns \emph{from},
\emph{to} (or POSIXct columns \emph{tBeg} and \emph{tEnd}) defining the
first and last timestamp of the time intervals in which the correlation is
to be used}

\item{hydraulicEvents}{hydraulic events}

\item{additionalColumns}{columns additional to "DateTime", "H" and "Q" to be
selected from \emph{hydraulicData}}

\item{modelDir}{full path to the directory where the model files (of the
current station) are stored)}

\item{settings}{settings as returned by \code{\link{configure}}. Will be used
to lookup function parameters for which no values have been given (see
defaults)}
}
\description{
1. remove rows with duplicate timestamps 2. fill gaps within hydraulic events
given by time intervals exceeding H threshold
}
