% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{whichAboveThresholds}
\alias{whichAboveThresholds}
\title{In Which Rows Are Thresholds Exceeded?}
\usage{
whichAboveThresholds(hydraulicData, indices = seq_len(nrow(hydraulicData)),
  thresholds = c(H = NA, Q = NA), columns = c(H = "H", Q = "Q"))
}
\arguments{
\item{hydraulicData}{data frame with columns as named in \code{columns}}

\item{indices}{vector of indices of preselected rows from which to exclude
those in which the thresholds are exceeded}

\item{thresholds}{vector of thresholds for \code{H} and \code{Q} values, 
respectively}

\item{columns}{vector of names containing \code{H} and \code{Q} values, 
respectively}
}
\description{
Get indices of rows in \emph{hydraulicData} in which H or Q thresholds are
  exceeded
}
