% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeCsvToPathFromDictionary}
\alias{writeCsvToPathFromDictionary}
\title{Write CSV to Path From Dictionary}
\usage{
writeCsvToPathFromDictionary(dataFrame, key, settings,
  open.directory = TRUE, ...)
}
\arguments{
\item{dataFrame}{data frame containing data to save}

\item{key}{key in \emph{settings$dictionary} to be resolved to file path}

\item{settings}{list of settings with elements \emph{dictionary},
\emph{outsep}, \emph{outdec}}

\item{open.directory}{if TRUE (default), the directory in which the file is
created is opened in the Windows Explorer after the file has been written.}

\item{\dots}{arguments passed to \code{\link{getOrCreatePath}}}
}
\description{
Write CSV to Path From Dictionary
}
