% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeDataAndOpenDirectory}
\alias{writeDataAndOpenDirectory}
\title{Write Data and Open Directory}
\usage{
writeDataAndOpenDirectory(dataFrame, filePath, settings)
}
\arguments{
\item{dataFrame}{data frame containing data to save}

\item{filePath}{base name of file ("_<moniPoint>_<sampleName>.csv" will be
appended)}

\item{settings}{list of settings, as returned by \code{\link{configure}}}
}
\description{
Write Data and Open Directory
}
